+++ src/components/FormSteper/index.js
@@ -1,75 +1,19 @@
-import { useState, Fragment, useContext } from 'react';
-import FormContext from '~/context/FormContext';
-import {
-  Box,
-  Stepper,
-  Step,
-  StepLabel,
-  Button,
-  Typography,
-} from '~/components';
+import { useContext } from 'react';
+import { FormContext } from '~/context/FormContext';
+import { Stepper, Step, StepLabel } from '~/components';
 
 const FormStepper = ({ steps }) => {
-  // const [activeStep, setActiveStep] = useState(0);
-  const { activeStep, setActiveStep } = useContext(FormContext);
-
-  const handleNext = () => {
-    setActiveStep((prevActiveStep) => prevActiveStep + 1);
-  };
-
-  const handleBack = () => {
-    setActiveStep((prevActiveStep) => prevActiveStep - 1);
-  };
-
-  const handleShowComponent = () => {
-    const { component } = steps.find((step) => step.index === activeStep);
-    return component;
-  };
+  const { activeStep } = useContext(FormContext);
   return (
-    <Box sx={{ width: '100%' }}>
-      <Stepper activeStep={activeStep}>
-        {steps.map(({ label }) => {
-          return (
-            <Step key={label}>
-              <StepLabel>{label}</StepLabel>
-            </Step>
-          );
-        })}
-      </Stepper>
-      {activeStep === steps.length ? (
-        <Fragment>
-          <Typography sx={{ mt: 2, mb: 1 }}>
-            All steps completed - you&apos;re finished
-          </Typography>
-          <Box sx={{ display: 'flex', flexDirection: 'row', pt: 2 }}>
-            <Box sx={{ flex: '1 1 auto' }} />
-          </Box>
-        </Fragment>
-      ) : (
-        <Fragment>
-          <Box>{handleShowComponent()}</Box>
-
-          <Box sx={{ display: 'flex', flexDirection: 'row', pt: 2 }}>
-            <Button
-              color="inherit"
-              disabled={activeStep === 0}
-              onClick={handleBack}
-              sx={{ mr: 1 }}
-            >
-              Back
-            </Button>
-            <Box sx={{ flex: '1 1 auto' }} />
-            {activeStep === steps.length - 1 ? (
-              <Button type="submit" variant="outlined">
-                Register
-              </Button>
-            ) : (
-              <Button onClick={handleNext}>Next</Button>
-            )}
-          </Box>
-        </Fragment>
-      )}
-    </Box>
+    <Stepper activeStep={activeStep}>
+      {steps.map(({ label }) => {
+        return (
+          <Step key={label}>
+            <StepLabel>{label}</StepLabel>
+          </Step>
+        );
+      })}
+    </Stepper>
   );
 };
