+++ src/context/FormContext.js
@@ -1,5 +1,16 @@
-import { createContext } from 'react';
+import { useState, createContext, useMemo } from 'react';
 
 const FormContext = createContext();
 
-export default FormContext;
+const FormProvider = ({ children }) => {
+  const [activeStep, setActiveStep] = useState(0);
+  const value = useMemo(() => {
+    return {
+      activeStep,
+      setActiveStep,
+    };
+  }, [activeStep]);
+  return <FormContext.Provider value={value}>{children}</FormContext.Provider>;
+};
+
+export { FormProvider, FormContext };
